+++ models/finances/expense/index.js
@@ -15,7 +15,7 @@ const financesExpense = connection.define('financesExpense', {
         allowNull: false
     },
     value: {
-        type: Sequelize.BIGINT,
+        type: Sequelize.DECIMAL(10, 2),
         allowNull: false
     },
     date: {
